bring all implemenation to calling view model
