+++ src/script/calling/CallingRepository.ts
@@ -124,8 +124,6 @@ enum CALL_DIRECTION {
 
 type SubconversationData = {epoch: number; secretKey: string; members: SubconversationEpochInfoMember[]};
 
-const MAX_USERS_TO_CALL_WITHOUT_CONFIRM = 4;
-
 export class CallingRepository {
   private readonly acceptVersionWarning: (conversationId: QualifiedId) => void;
   private readonly callLog: string[];
@@ -514,7 +512,6 @@ export class CallingRepository {
    */
   subscribeToEvents(): void {
     amplify.subscribe(WebAppEvents.CALL.EVENT_FROM_BACKEND, this.onCallEvent);
-    amplify.subscribe(WebAppEvents.CALL.STATE.TOGGLE, this.toggleState); // This event needs to be kept, it is sent by the wrapper
     amplify.subscribe(WebAppEvents.PROPERTIES.UPDATE.CALL.ENABLE_VBR_ENCODING, this.toggleCbrEncoding);
     amplify.subscribe(WebAppEvents.PROPERTIES.UPDATED, ({settings}: WebappProperties) => {
       this.toggleCbrEncoding(settings.call.enable_vbr_encoding);
@@ -684,42 +681,6 @@ export class CallingRepository {
   //##############################################################################
   // Call actions
   //##############################################################################
-
-  private readonly toggleState = async (withVideo: boolean): Promise<void> => {
-    const conversation = this.conversationState.activeConversation();
-    if (conversation) {
-      const isActiveCall = this.findCall(conversation.qualifiedId);
-      const callType = withVideo ? CALL_TYPE.VIDEO : CALL_TYPE.NORMAL;
-
-      if (isActiveCall) {
-        this.leaveCall(conversation.qualifiedId, LEAVE_CALL_REASON.ELECTRON_TRAY_MENU_MESSAGE);
-        return;
-      }
-
-      const memberCount = conversation.participating_user_ets().length;
-
-      if (memberCount > MAX_USERS_TO_CALL_WITHOUT_CONFIRM) {
-        PrimaryModal.show(PrimaryModal.type.WITHOUT_TITLE, {
-          preventClose: true,
-          primaryAction: {
-            action: async () => await this.startCall(conversation, callType),
-            text: t('groupCallModalPrimaryBtnName'),
-          },
-          secondaryAction: {
-            text: t('modalConfirmSecondary'),
-          },
-          text: {
-            htmlMessage: `<div class="modal-description">
-            ${t('groupCallConfirmationModalTitle', memberCount)}
-          </div>`,
-            closeBtnLabel: t('groupCallModalCloseBtnLabel'),
-          },
-        });
-      } else {
-        await this.startCall(conversation, callType);
-      }
-    }
-  };
 
   private getConversationType(conversation: Conversation): CONV_TYPE {
     if (!conversation.isGroup()) {
